requirements.txt
tk
pdf2image
numpy
pillow
img2pdf
